tf-service/requirements.txt
fastapi
uvicorn
pytest
pytest-asyncio
httpx
python-multipart
tensorflow
opencv-python-headless
numpy
python-dotenv